refactor(api): move the Kakao geocode contract into a server adapter
The route knew the whole upstream contract: the API key, the URL, the
KakaoAK header, the non-JSON guard, and that Kakao reports failures in
a 200 body via errorType rather than only through the status code. That
left the Kakao contract living under src/app/api, so a change on their
side would have been fixed there.

geocodeAddress moves it to adapters/server/kakao/ — named for the
service, mirroring the existing browser/kakao/ loader, since server/ and
browser/ already carry the runtime distinction.

The segment becomes /api/kakao-map, following the compound-segment rule
added in #233. Only the route path changes: the schema file, the
KakaomapResponse type and useKakaomapGeocode keep their names, because
that rule governs route segments. The kakaomap:// deeplink stays too —
that scheme is Kakao's, not ours.

The address was interpolated into the upstream query string raw, so an
address containing & or # could truncate the request or inject
parameters. The adapter encodes it.

The five route tests move with the segment and pass with only their URL
strings changed, which is the evidence that behaviour is preserved.
